# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hariri Foundation API",
        "description": "API documentation for Users Management",
        "contact": {
            "email": "support@haririfoundation.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/surveys": {
            "get": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Get all surveys or a specific survey by ID",
                "operationId": "ee59325f71cde867ed089dba29975752",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Optional survey UUID to fetch a specific survey",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Surveys retrieved successfully"
                    },
                    "404": {
                        "description": "Survey not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Create a new survey",
                "operationId": "e56eadc7ab1db57890075dc84c9ff369",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "End of Year Evaluation"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://example.com/survey-link"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Survey created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/surveys/{id}": {
            "put": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Update an existing survey",
                "operationId": "eea04041ecc6b9bfaa996a9e84f05ae6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Survey UUID to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Updated survey description"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://new-link.com"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Survey updated successfully"
                    },
                    "404": {
                        "description": "Survey not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Surveys"
                ],
                "summary": "Delete a survey by UUID",
                "operationId": "ef02b30c17f8357878d30403e82aca1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Survey UUID to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Survey deleted successfully"
                    },
                    "404": {
                        "description": "Survey not found"
                    }
                }
            }
        },
...
    },
...
}